Confirmed accuracy of update with test search method
Added a test search method for the client that creates a search
query the same way the 'search' method will, and tested it against
the index created by 'update'. Works, although not exhaustively
tested.  Great news, meaning we can go ahead with implemented the
server-side of the SSE routines with confidence.

Additionally changed some stuff around in the client and server's
structure, making the backends the engine and the network elements
part of the engines rather than the other way around (at least for
the client. I don't think I changed that in the server, though I
will in an upcoming commit.