Completed Update SSE routine implementation
Update is now implemented in both client and server. Existing
entries do not get overwritten. Have to do a little extra
computation to encode the JSON on the server side to ascii
from unicode, although it could just be a packaging issue on
the client side. Will try to sort that out to avoid the
extra library calls.