Completed SSE Search implementation on Server (minus sending the documents back to client)
Server recieves k1 and k2 from client, computes HMAC to find correct
index (or indices) for the query, and decrypts the documents names
of the docs that contain the query.

Client was actually hashing document names instead of encrypting
them, so that was updated as well.

Server doesn't yet send back the actual documents, but that will be
an easy addition.