Added some semblance of file transfer to the SEARCH routine
Now files that are found to contain the search query are sent
to client. However, there isn't any good logic/infrastructure
in place to know which files to write the data to (since the
files are sent as an arbitrary data stream from the server,
the client doesn't know which file it is).  Currently, it just
writes (and subsequently overwrites) a single file on the client
side.

It's also only one-way. The client does not send the server new
messages when it runs the UPDATE routine.

Not a good solution, but it's something to work with for the time
being. Also retested UPDATE with the new SEARCH routine, and added
new files to the server index successfully (but, as mentioned, the
files themselves don't make it over to the server).

First order of business is sorting out a better solution for file
transfer, then to bring BOTH client and server up-to-speed.