Added beginnings of proper email (MIME) message usage
Client now reads in MIME messages, parses the body and generates
index.  For now, only body, but parsing the headers is easy and
next in line.  The only confusion will be how to store them so
users can search by header.  Tested with UPDATE and SEARCH, MIME
messages are from Enron email corpus.

Additionally, did a little clean up.  Changed the way client was
encrypting and sending the document names in the index. Was very
dependent on how the server stored the messages. Now less so.