Added basic header parsing and indexing for SSE
Now headers are added to index, but in a very simple and naive
way.  No stripping occurs, and no stemming, so search terms in
relation to headers are a bit opaque.  Will revisit how I want
that to work in a later commit.

Also added proper handling of SEARCH failures (ie: if server
doesn't find search query).