Added UPDATE and storage of email headers
Headers for MIME messages now get added both to the index as
normal text would be, but also get additionally added PRF'd
with their header type.  This means they'll be stored again
under a new hashed value that can be evaluated by the server
in a similar way (PRF with header type instead of count #).

Additionally, made a change or two to the original UPDATE routine
that should cut out some unecessary iteration, and added a bunch
of commnents in an attempt to prepare documentation.