Implemented basic SEARCH on headers
Not very clean, and taking on some technical debt, as I'll need
(or at least want) to clean it up.  Sloppy, but gotta get this
done to carry on.  Added args to some routines to manage search,
and added a function or two. None of the changes are particularly
worthy, but will have to do for now.

Tested with updating and searching for a header term.