Implemented new index scheme for UPDATE on client.
New scheme (prototype to test against the master branch's
performance) now creates a list of message ids that are encrypted,
so when the server is searching, it can decrypt all the id's in
one step, rather than decrypt separately for each occurance of
the query in as many files as it appears.

Note that as of this commit, it has only been implemented on the
client, and while it is successful executing UPDATE, it will fail
or return unexpected results for SEARCH.  The server needs to know
nothing of the changes regarding update, and inversely, only the
server needs to know about the new implementation in regards to
SEARCH, which will follow in the next commit or two.