Fixed UPDATE issues.
My mistake. Previous checkin didn't have the required server-side
implementation (which I thought wasn't needed). Actually ran into
some issues so it required more effort than expected...

It's difficult to determine if a PRF'd key value (say, in the
server's index) is the value of the new to-be-added key with a
smaller id list.  Eventually, I decided to just send along another
value (lprime in the code for now) that is the PRF'd version of the
new term || c-1.  This may mean more leakage, as a listener will
get both your update term, and your update term -1.  Had to get it
sorted out now though, so it'll be revisited.

Now to implement SEARCH to see if this will work at all.