Added SEARCH routine to SSE server
The general idea (and it's kinda a work-around for now) is that the
client sends k1 and k2 (as usual), but also kprime, which is the PRF
of w || c-1, where as k1 is w || c.  That way, the server can quickly
identify and overwrite the original entry with the shorter id list.
This could lead to leakage, but will have to be sorted out later.

It works, but it seems really slow. Could be that I'm testing with
more files (and therefore a larger index) than when I tested the
original implementation on master (1 encrypted id per k value).