Updated algorithms for SEARCH and UPDATE
Massively improves SEARCH efficiency at the cost of revealing 'c'
(count of documents that term is found in) in transit and to server.
Server can find out 'c' anyway, but it is visible on the wire.
Shouldn't be a huge issue, especially because each update will alter
that key (representing the term) in the index.

Made some other, simpler changes to the algorithms that improve
performance, but not on the order as that mentioned above.

Additionally, updated time.sh to include command-line args.